Updating video display page
