+++ client/src/App.jsx
@@ -22,7 +22,6 @@ import TeacherLogin from './views/TeacherLogin/TeacherLogin';
 import VideoEditor from './views/ContentCreator/VideoEditor/VideoEditor';
 import VideoDisplay from './views/ContentCreator/VideoEditor/VideoDisplay';
 import VideoPage from './views/Video/VideoPage';
-import TestVideo from './views/Video/TestVideo';
 
 const App = () => {
   return (
@@ -35,11 +34,8 @@ const App = () => {
         <Route path='/reset-password' element={<ResetPassword />} />
         <Route path='/login' element={<StudentLogin />} />
         <Route path='/replay/:saveID' element={<Replay />} />
-        <Route path='video-player' element={<VideoPage />} />
-        <Route path='test-video' element={<TestVideo />} />
         <Route path='/sandbox' element={<BlocklyPage isSandbox={true} />} />
-        <Route path='video-player' element={<VideoPage />} />
-        <Route path='test-video' element={<TestVideo />} />
+        <Route path='/video' element={<VideoPage />} />
         <Route
           path='/report'
           element={
